pyproject.toml
[project]
name = "llm-data-pipeline"
version = "0.1.0"
description = "Ray + asyncio pipeline for LLM batch processing with resume support."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "azure-identity>=1.17.1",
  "httpx>=0.27.0",
  "orjson>=3.10.0",
  "pydantic>=2.6.0",
  "ray[default]>=2.9.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["pipeline*"]
